API Reference

model.jl

DecisionProgramming.Node — Type

Node type.

DecisionProgramming.State — Type

State type.

DecisionProgramming.InfluenceDiagram — Type

Influence diagram.

DecisionProgramming.Probabilities — Type

Probabilities type.

DecisionProgramming.Consequences — Type

Consequences type.

DecisionProgramming.Path — Type

Path type.

DecisionProgramming.UtilityFunction — Type

UtilityFunction type. Maps Path to Real.

DecisionProgramming.DecisionModel — Type

Defines the DecisionModel type.

DecisionProgramming.DecisionStrategy — Type

Decision strategy type.

DecisionProgramming.paths — Function

Iterate over paths.

Iterate over paths with fixed states.

DecisionProgramming.path_probability — Function

Path probability (upper bound).

DecisionProgramming.InfluenceDiagram — Method

Construct and validate an influence diagram.

Arguments

C::Vector{Node}: Change nodes.

D::Vector{Node}: Decision nodes.

V::Vector{Node}: Value nodes.

A::Vector{Pair{Node, Node}}: Arcs between nodes.

S_j::Vector{State}: Number of states.

DecisionProgramming.validate_probabilities — Method

Validate probabilities.

DecisionProgramming.validate_consequences — Method

Validate consequences.

DecisionProgramming.variables — Function

Create a multidimensional array of variables.

DecisionProgramming.DecisionModel — Method

Construct a DecisionModel from an influence diagram and parameters.

Arguments

G::InfluenceDiagram

X::Probabilities

positive_path_utility::Bool=true

DecisionProgramming.probability_sum_cut — Function

Adds a probability sum cut to the model as a lazy constraint.

DecisionProgramming.number_of_paths_cut — Function

Adds a number of paths cut to the model as a lazy constraint.

DecisionProgramming.transform_affine_positive — Function

Affine positive tranformation of path utility.

DecisionProgramming.expected_value — Function

Expected value.

DecisionProgramming.conditional_value_at_risk — Function

Conditional value-at-risk (CVaR).

DecisionProgramming.DecisionStrategy — Method

Extract values for decision variables from a decision model.

analysis.jl

DecisionProgramming.ActivePaths — Type

Interface for iterating over active paths given influence diagram and decision strategy.

for s in ActivePaths(G, Z)
    ...
end

DecisionProgramming.utility_distribution — Function

The probability mass function.

DecisionProgramming.state_probabilities — Function

State probabilities.

State probabilities.

printing.jl

DecisionProgramming.print_decision_strategy — Function

Print decision strategy.

DecisionProgramming.print_state_probabilities — Function

Print state probabilities with fixed states.

Print state probabilities.

random.jl

DecisionProgramming.random_influence_diagram — Function

Create random influence diagram of given size.

DecisionProgramming.random_probabilities — Function

Generate random probabilities

DecisionProgramming.random_consequences — Function

Generate random utilities