API Reference

model.jl

DecisionProgramming.paths — Function

Iterate over paths.

Iterate over paths with fixed states.

DecisionProgramming.path_probability — Function

Path probability (upper bound).

DecisionProgramming.InfluenceDiagram — Type

Influence diagram.

DecisionProgramming.Probabilities — Type

Probabilities type.

DecisionProgramming.Consequences — Type

Consequences type.

DecisionProgramming.UtilityFunction — Type

UtilityFunction type. Maps path to real.

DecisionProgramming.DecisionModel — Type

Defines the DecisionModel type.

DecisionProgramming.DecisionStrategy — Type

Decision strategy type.

DecisionProgramming.InfluenceDiagram — Method

Construct and validate an influence diagram.

Arguments

C::Vector{Int}: Change nodes.

D::Vector{Int}: Decision nodes.

V::Vector{Int}: Value nodes.

A::Vector{Pair{Int, Int}}: Arcs between nodes.

S_j::Vector{Int}: Number of states.

DecisionProgramming.validate_probabilities — Method

Validate probabilities.

DecisionProgramming.validate_consequences — Method

Validate consequences.

DecisionProgramming.DecisionModel — Method

Construct a DecisionModel from an influence diagram and parameters.

Arguments

G::InfluenceDiagram

X::Probabilities

positive_path_utility::Bool=true

DecisionProgramming.probability_sum_cut — Function

Adds a probability sum cut to the model as a lazy constraint.

DecisionProgramming.number_of_paths_cut — Function

Adds a number of paths cut to the model as a lazy constraint.

DecisionProgramming.transform_affine_positive — Function

Affine positive tranformation of path utility.

DecisionProgramming.expected_value — Function

Expected value.

DecisionProgramming.value_at_risk — Function

Value-at-risk.

DecisionProgramming.DecisionStrategy — Method

Extract values for decision variables from a decision model.

analysis.jl

DecisionProgramming.is_compatible — Function

Test is path is compatible with a decision strategy.

DecisionProgramming.active_paths — Function

Generate all active paths from a decision strategy with fixed states.

Generate all active paths from a decision strategy.

DecisionProgramming.utility_distribution — Function

The probability mass function.

DecisionProgramming.state_probabilities — Function

State probabilities.

State probabilities.

printing.jl

DecisionProgramming.print_decision_strategy — Function

Print decision strategy.

DecisionProgramming.print_state_probabilities — Function

Print state probabilities with fixed states.

Print state probabilities.

random.jl

DecisionProgramming.random_influence_diagram — Function

Create random influence diagram of given size.

DecisionProgramming.random_probabilities — Function

Generate random probabilities

DecisionProgramming.random_consequences — Function

Generate random utilities