API Reference

DecisionProgramming.jl API reference.

model.jl

Types

DecisionProgramming.Node — Type

Node type.

DecisionProgramming.State — Type

State type.

DecisionProgramming.InfluenceDiagram — Type

Influence diagram.

DecisionProgramming.Probabilities — Type

Probabilities type.

DecisionProgramming.Consequences — Type

Consequences type.

DecisionProgramming.DecisionStrategy — Type

Decision strategy type.

DecisionProgramming.Path — Type

Path type.

DecisionProgramming.PathUtility — Type

PathUtility type. Path utility is a function that maps paths to real values.

Examples

U(s::Path)::Real = ...

DecisionProgramming.DecisionModel — Type

DecisionModel type.

Path Functions

DecisionProgramming.paths — Function

Iterate over paths in lexicographical order.

Examples

julia> collect(paths([2, 3]))
(1, 1), (1, 2), (1, 3), (2, 1), (2, 2), (2, 3)

Iterate over paths with fixed states in lexicographical order.

Examples

julia> collect(paths([2, 3], fixed=Dict(1=>2)))
(2, 1), (2, 2), (2, 3)

DecisionProgramming.path_probability — Function

Path probability (upper bound).

Influence Diagram

DecisionProgramming.InfluenceDiagram — Method

Construct and validate an influence diagram.

Examples

C = [1, 3]
D = [2, 4]
V = [5]
A = [1=>2, 3=>4, 2=>5, 3=>5]
S_j = [2, 3, 2, 4]
G = InfluenceDiagram(C, D, V, A, S_j)

DecisionProgramming.validate_probabilities — Method

Validate probabilities for an influence diagram.

Examples

X = ...  # Unvalidated probabilities
X = validate_probabilities(G, X)

DecisionProgramming.validate_consequences — Method

Validate consequences for an influence diagram.

Examples

Y = ...  # Unvalidated consequences
Y = validate_consequences(G, Y)

Decision Model

DecisionProgramming.DecisionModel — Method

Construct a DecisionModel from an influence diagram and probabilities.

Examples

model = DecisionModel(G, X; positive_path_utility=true)

DecisionProgramming.probability_sum_cut — Function

Adds a probability sum cut to the model as a lazy constraint.

Examples

probability_sum_cut(model, G, X)

DecisionProgramming.number_of_paths_cut — Function

Adds a number of paths cut to the model as a lazy constraint.

Examples

atol = 0.9  # Tolerance to trigger the creation of the lazy cut
number_of_paths_cut(model, G, X; atol=atol)

Objective Functions

DecisionProgramming.transform_affine_positive — Function

Affine positive tranformation of path utility.

Examples

U(s::Path)::Real = ...
U⁺ = transform_affine_positive(G, U)

DecisionProgramming.expected_value — Function

Expected value objective.

Examples

EV = expected_value(model, G, U⁺)

DecisionProgramming.conditional_value_at_risk — Function

Conditional value-at-risk (CVaR) objective. Also known as Expected Shortfall (ES).

Examples

α = 0.05  # Parameter such that 0 ≤ α ≤ 1
ES = conditional_value_at_risk(model, G, U⁺, α)

Decision Strategy

DecisionProgramming.DecisionStrategy — Method

Extract values for decision variables from a decision model.

Examples

Z = DecisionStrategy(model)

analysis.jl

DecisionProgramming.ActivePaths — Type

Interface for iterating over active paths given influence diagram and decision strategy.

Initialize path s of length n

Fill chance states s[C] by generating subpaths paths(G.C)

Fill decision states s[D] by decision strategy Z and path s

Examples

for s in ActivePaths(G, Z)
    ...
end

DecisionProgramming.UtilityDistribution — Type

UtilityDistribution type.

DecisionProgramming.UtilityDistribution — Method

Constructs the probability mass function for path utilities on active paths.

Examples

UtilityDistribution(G, X, Z, U)

DecisionProgramming.StateProbabilities — Type

StateProbabilities type.

DecisionProgramming.StateProbabilities — Method

Associates each node with array of probabilities for each of its states occuring in active paths.

Examples

StateProbabilities(G, X, Z)

DecisionProgramming.StateProbabilities — Method

Associates each node with array of conditional probabilities for each of its states occuring in active paths given fixed states and prior probability.

Examples

# Prior probabilities
prev = StateProbabilities(G, X, Z)

# Select node and fix its state
node = 1
state = 2
StateProbabilities(G, X, Z, node, state, prev)

printing.jl

DecisionProgramming.print_decision_strategy — Function

Print decision strategy.

Examples

print_decision_strategy(G, Z)

DecisionProgramming.print_state_probabilities — Function

Print state probabilities with fixed states.

Examples

sprobs = StateProbabilities(G, X, Z)
print_state_probabilities(sprobs, [1, 2], ["A", "B"])

random.jl

DecisionProgramming.random_influence_diagram — Function

Create random influence diagram of given size.

Examples

n_C, n_D, n_V = 4, 2, 1  # Number of chance, decision and value states
n_A = 2  # Upper limit on information set size for any node
num_states = [2, 3]  # Set of possible number of states
G = random_influence_diagram(n_C, n_D, n_V, n_A, num_states)

DecisionProgramming.random_probabilities — Function

Generate random probabilities for an influence diagram.

Examples

X = random_probabilities(G)

DecisionProgramming.random_consequences — Function

Generate random consequences for an influence diagram.

Examples

Y = random_consequences(G)