API Reference

DecisionProgramming.jl API reference.

influence_diagram.jl

Nodes

DecisionProgramming.Node — Type

Node = Int

Primitive type for node. Alias for Int.

DecisionProgramming.ChanceNode — Type

ChanceNode <: AbstractNode

Chance node type.

Examples

c = ChanceNode(3, [1, 2])

DecisionProgramming.DecisionNode — Type

DecisionNode <: AbstractNode

Decision node type.

Examples

d = DecisionNode(2, [1])

DecisionProgramming.ValueNode — Type

ValueNode <: AbstractNode

Value node type.

Examples

v = ValueNode(4, [1, 3])

DecisionProgramming.State — Type

const State = Int

Primitive type for the number of states. Alias for Int.

DecisionProgramming.States — Type

States <: AbstractArray{State, 1}

States type. Works like Vector{State}.

Examples

S = States([2, 3, 2, 4])

DecisionProgramming.States — Method

function States(states::Vector{Tuple{State, Vector{Node}}})

Construct states from vector of (state, nodes) tuples.

Examples

julia> S = States([(2, [1, 3]), (3, [2, 4, 5])])
States([2, 3, 2, 3, 3])

DecisionProgramming.validate_influence_diagram — Function

function validate_influence_diagram(S::States, C::Vector{ChanceNode}, D::Vector{DecisionNode}, V::Vector{ValueNode})

Validate influence diagram.

Paths

DecisionProgramming.Path — Type

const Path{N} = NTuple{N, State} where N

Path type. Alias for NTuple{N, State} where N.

DecisionProgramming.paths — Method

function paths(states::AbstractVector{State})

Iterate over paths in lexicographical order.

Examples

julia> states = States([2, 3])
julia> vec(collect(paths(states)))
[(1, 1), (2, 1), (1, 2), (2, 2), (1, 3), (2, 3)]

DecisionProgramming.paths — Method

function paths(states::AbstractVector{State}, fixed::Dict{Node, State})

Iterate over paths with fixed states in lexicographical order.

Examples

julia> states = States([2, 3])
julia> vec(collect(paths(states, Dict(1=>2))))
[(2, 1), (2, 2), (2, 3)]

DecisionProgramming.ForbiddenPath — Type

const ForbiddenPath = Tuple{Vector{Node}, Set{Path}}

ForbiddenPath type.

Examples

ForbiddenPath[
    ([1, 2], Set([(1, 2)])),
    ([3, 4, 5], Set([(1, 2, 3), (3, 4, 5)]))
]

Probabilities

DecisionProgramming.Probabilities — Type

struct Probabilities{N} <: AbstractArray{Float64, N}

Construct and validate stage probabilities.

Examples

julia> data = [0.5 0.5 ; 0.2 0.8]
julia> X = Probabilities(2, data)
julia> s = (1, 2)
julia> X(s)
0.5

Path Probability

DecisionProgramming.AbstractPathProbability — Type

abstract type AbstractPathProbability end

Abstract path probability type.

Examples

struct PathProbability <: AbstractPathProbability
    C::Vector{ChanceNode}
    # ...
end

(U::PathProbability)(s::Path) = ...

DecisionProgramming.DefaultPathProbability — Type

DefaultPathProbability <: AbstractPathProbability

Path probability.

Examples

P = DefaultPathProbability(C, X)
s = (1, 2)
P(s)

Consequences

DecisionProgramming.Consequences — Type

Consequences{N} <: AbstractArray{Float64, N}

State utilities.

Examples

julia> vals = [1.0 -2.0; 3.0 4.0]
julia> Y = Consequences(3, vals)
julia> s = (1, 2)
julia> Y(s)
-2.0

Path Utility

DecisionProgramming.AbstractPathUtility — Type

abstract type AbstractPathUtility end

Abstract path utility type.

Examples

struct PathUtility <: AbstractPathUtility
    V::Vector{ValueNode}
    # ...
end

(U::PathUtility)(s::Path) = ...

DecisionProgramming.DefaultPathUtility — Type

DefaultPathUtility <: AbstractPathUtility

Default path utility.

Examples

U = DefaultPathUtility(V, Y)
s = (1, 2)
U(s)

Decision Strategy

DecisionProgramming.LocalDecisionStrategy — Type

LocalDecisionStrategy{N} <: AbstractArray{Int, N}

Local decision strategy type.

Examples

Z = LocalDecisionStrategy(1, data)
Z(s_I)

DecisionProgramming.DecisionStrategy — Type

DecisionStrategy

Decision strategy type.

decision_model.jl

Decision Model

DecisionProgramming.DecisionVariables — Type

DecisionVariables(model::Model, S::States, D::Vector{DecisionNode}; names::Bool=false, name::String="z")

Create decision variables and constraints.

Arguments

model::Model: JuMP model into which variables are added.

S::States: States structure associated with the influence diagram.

D::Vector{DecisionNode}: Vector containing decicion nodes.

names::Bool: Use names or have JuMP variables be anonymous.

name::String: Prefix for predefined decision variable naming convention.

Examples

z = DecisionVariables(model, S, D)

DecisionProgramming.PathCompatibilityVariables — Type

PathCompatibilityVariables(model::Model,
    z::DecisionVariables,
    S::States,
    P::AbstractPathProbability;
    names::Bool=false,
    name::String="x",
    forbidden_paths::Vector{ForbiddenPath}=ForbiddenPath[],
    fixed::Dict{Node, State}=Dict{Node, State}(),
    probability_cut::Bool=true)

Create path compatibility variables and constraints.

Arguments

model::Model: JuMP model into which variables are added.

z::DecisionVariables: Decision variables from DecisionVariables function.

S::States: States structure associated with the influence diagram.

P::AbstractPathProbability: Path probabilities structure for which the function P(s) is defined and returns the path probabilities for path s.

names::Bool: Use names or have JuMP variables be anonymous.

name::String: Prefix for predefined decision variable naming convention.

forbidden_paths::Vector{ForbiddenPath}: The forbidden subpath structures. Path compatibility variables will not be generated for paths that include forbidden subpaths.

fixed::Dict{Node, State}: Path compatibility variable will not be generated for paths which do not include these fixed subpaths.

probability_cut Includes probability cut constraint in the optimisation model.

Examples

x_s = PathCompatibilityVariables(model, z, S, P; probability_cut = false)

DecisionProgramming.lazy_probability_cut — Method

lazy_probability_cut(model::Model, x_s::PathCompatibilityVariables, P::AbstractPathProbability)

Adds a probability cut to the model as a lazy constraint.

Examples

lazy_probability_cut(model, x_s, P)

Note

Remember to set lazy constraints on in the solver parameters, unless your solver does this automatically. Note that Gurobi does this automatically.

Objective Functions

DecisionProgramming.PositivePathUtility — Type

PositivePathUtility(S::States, U::AbstractPathUtility)

Positive affine transformation of path utility. Always evaluates positive values.

Examples

julia> U⁺ = PositivePathUtility(S, U)
julia> all(U⁺(s) > 0 for s in paths(S))
true

DecisionProgramming.NegativePathUtility — Type

NegativePathUtility(S::States, U::AbstractPathUtility)

Negative affine transformation of path utility. Always evaluates negative values.

Examples

julia> U⁻ = NegativePathUtility(S, U)
julia> all(U⁻(s) < 0 for s in paths(S))
true

DecisionProgramming.expected_value — Method

expected_value(model::Model,
    x_s::PathCompatibilityVariables,
    U::AbstractPathUtility,
    P::AbstractPathProbability;
    probability_scale_factor::Float64=1.0)

Create an expected value objective.

Arguments

model::Model: JuMP model into which variables are added.

x_s::PathCompatibilityVariables: Path compatibility variables.

S::States: States structure associated with the influence diagram.

P::AbstractPathProbability: Path probabilities structure for which the function P(s) is defined and returns the path probabilities for path s.

probability_scale_factor::Float64: Multiplies the path probabilities by this factor.

Examples

EV = expected_value(model, x_s, U, P)
EV = expected_value(model, x_s, U, P; probability_scale_factor = 10.0)

DecisionProgramming.conditional_value_at_risk — Method

conditional_value_at_risk(model::Model,
    x_s::PathCompatibilityVariables{N},
    U::AbstractPathUtility,
    P::AbstractPathProbability,
    α::Float64;
    probability_scale_factor::Float64=1.0) where N

Create a conditional value-at-risk (CVaR) objective.

Arguments

model::Model: JuMP model into which variables are added.

x_s::PathCompatibilityVariables: Path compatibility variables.

S::States: States structure associated with the influence diagram.

P::AbstractPathProbability: Path probabilities structure for which the function P(s) is defined and returns the path probabilities for path s.

α::Float64: Probability level at which conditional value-at-risk is optimised.

probability_scale_factor::Float64: Adjusts conditional value at risk model to be compatible with the expected value expression if the probabilities were scaled there.

Examples

α = 0.05  # Parameter such that 0 ≤ α ≤ 1
CVaR = conditional_value_at_risk(model, x_s, U, P, α)
CVaR = conditional_value_at_risk(model, x_s, U, P, α; probability_scale_factor = 10.0)

Decision Strategy from Variables

DecisionProgramming.LocalDecisionStrategy — Method

LocalDecisionStrategy(j::Node, z::Array{VariableRef})

Construct decision strategy from variable refs.

DecisionProgramming.DecisionStrategy — Method

DecisionStrategy(z::DecisionVariables)

Extract values for decision variables from solved decision model.

Examples

Z = DecisionStrategy(z)

analysis.jl

Missing docstring.

Missing docstring for CompatiblePaths. Check Documenter's build log for details.

DecisionProgramming.UtilityDistribution — Type

 UtilityDistribution

UtilityDistribution type.

DecisionProgramming.UtilityDistribution — Method

UtilityDistribution(S::States, P::AbstractPathProbability, U::AbstractPathUtility, Z::DecisionStrategy)

Constructs the probability mass function for path utilities on paths that are compatible and active.

Examples

UtilityDistribution(S, P, U, Z)

DecisionProgramming.StateProbabilities — Type

StateProbabilities

StateProbabilities type.

DecisionProgramming.StateProbabilities — Method

function StateProbabilities(S::States, P::AbstractPathProbability, Z::DecisionStrategy)

Associates each node with array of probabilities for each of its states occuring in active paths.

Examples

StateProbabilities(S, P, Z)

DecisionProgramming.StateProbabilities — Method

function StateProbabilities(S::States, P::AbstractPathProbability, Z::DecisionStrategy, node::Node, state::State, prev::StateProbabilities)

Associates each node with array of conditional probabilities for each of its states occuring in active paths given fixed states and prior probability.

Examples

# Prior probabilities
prev = StateProbabilities(S, P, Z)

# Select node and fix its state
node = 1
state = 2
StateProbabilities(S, P, Z, node, state, prev)

DecisionProgramming.value_at_risk — Method

function value_at_risk(u::Vector{Float64}, p::Vector{Float64}, α::Float64)

Value-at-risk.

DecisionProgramming.conditional_value_at_risk — Method

function conditional_value_at_risk(u::Vector{Float64}, p::Vector{Float64}, α::Float64)

Conditional value-at-risk.

printing.jl

DecisionProgramming.print_decision_strategy — Function

function print_decision_strategy(S::States, Z::DecisionStrategy)

Print decision strategy.

Examples

print_decision_strategy(S, Z)

DecisionProgramming.print_utility_distribution — Function

function print_utility_distribution(udist::UtilityDistribution; util_fmt="%f", prob_fmt="%f")

Print utility distribution.

Examples

udist = UtilityDistribution(S, P, U, Z)
print_utility_distribution(udist)

DecisionProgramming.print_state_probabilities — Function

function print_state_probabilities(sprobs::StateProbabilities, nodes::Vector{Node}; prob_fmt="%f")

Print state probabilities with fixed states.

Examples

sprobs = StateProbabilities(S, P, U, Z)
print_state_probabilities(sprobs, [c.j for c in C])
print_state_probabilities(sprobs, [d.j for d in D])

DecisionProgramming.print_statistics — Function

function print_statistics(udist::UtilityDistribution; fmt = "%f")

Print statistics about utility distribution.

DecisionProgramming.print_risk_measures — Function

function print_risk_measures(udist::UtilityDistribution, αs::Vector{Float64}; fmt = "%f")

Print risk measures.

random.jl

DecisionProgramming.random_diagram — Method

function random_diagram(rng::AbstractRNG, n_C::Int, n_D::Int, n_V::Int, m_C::Int, m_D::Int)

Generate random decision diagram with n_C chance nodes, n_D decision nodes, and n_V value nodes. Parameter m_C and m_D are the upper bounds for the size of the information set.

Examples

rng = MersenneTwister(3)
random_diagram(rng, 5, 2, 3, 2)

DecisionProgramming.States — Method

function States(rng::AbstractRNG, states::Vector{State}, n::Int)

Generate n random states from states.

Examples

rng = MersenneTwister(3)
S = States(rng, [2, 3], 10)

DecisionProgramming.Probabilities — Method

function Probabilities(rng::AbstractRNG, c::ChanceNode, S::States; n_inactive::Int=0)

Generate random probabilities for chance node c with S states.

Examples

rng = MersenneTwister(3)
c = ChanceNode(2, [1])
S = States([2, 2])
Probabilities(rng, c, S)

DecisionProgramming.Consequences — Method

function Consequences(rng::AbstractRNG, v::ValueNode, S::States; low::Float64=-1.0, high::Float64=1.0)

Generate random consequences between low and high for value node v with S states.

Examples

rng = MersenneTwister(3)
v = ValueNode(3, [1])
S = States([2, 2])
Consequences(rng, v, S; low=-1.0, high=1.0)

DecisionProgramming.LocalDecisionStrategy — Method

function LocalDecisionStrategy(rng::AbstractRNG, d::DecisionNode, S::States)

Generate random decision strategy for decision node d with S states.

Examples

rng = MersenneTwister(3)
d = DecisionNode(2, [1])
S = States([2, 2])
LocalDecisionStrategy(rng, d, S)