API

Model

DecisionProgramming.paths — Function

Iterate over paths.

Iterate over paths with fixed states.

DecisionProgramming.Specs — Type

Specification for different model scenarios. For example, we can specify toggling on and off constraints and objectives.

Arguments

probability_sum_cut::Bool: Toggle probability sum cuts on and off.

num_paths::Int: If larger than zero, enables the number of paths cuts using the supplied value.

DecisionProgramming.InfluenceDiagram — Type

Influence diagram.

DecisionProgramming.InfluenceDiagram — Method

Construct and validate an influence diagram.

Arguments

C::Vector{Int}: Change nodes.

D::Vector{Int}: Decision nodes.

V::Vector{Int}: Value nodes.

A::Vector{Pair{Int, Int}}: Arcs between nodes.

S_j::Vector{Int}: Number of states.

DecisionProgramming.Params — Type

Decision model parameters.

DecisionProgramming.Params — Method

Construct and validate decision model parameters.

Arguments

diagram::InfluenceDiagram: The influence diagram associated with the probabilities and consequences.

X::Dict{Int, Array{Float64}}: Probabilities

Y::Dict{Int, Array{Float64}}: Consequences

DecisionProgramming.DecisionModel — Type

Defines the DecisionModel type.

DecisionProgramming.DecisionModel — Method

Construct a DecisionModel from specification, influence diagram and parameters.

Arguments

specs::Specs

diagram::InfluenceDiagram

params::Params

DecisionProgramming.probability_sum_cut — Function

Probability sum lazy cut.

DecisionProgramming.number_of_paths_cut — Function

Number of paths lazy cut.

Analysis

DecisionProgramming.state_probabilities — Function

State probabilities.

State probabilities.