API

DecisionProgramming.Specs — Type

Specification for different model scenarios. For example, we can specify toggling on and off constraints and objectives.

Arguments

probability_sum_cut::Bool: Toggle probability sum cuts on and off.

num_paths::Int: If larger than zero, enables the number of paths cuts using the supplied value.

DecisionProgramming.DecisionGraph — Type

Decision graph is a directed, acyclic graph.

Arguments

C::Vector{Int}: Change nodes.

D::Vector{Int}: Decision nodes.

V::Vector{Int}: Value nodes.

A::Vector{Pair{Int, Int}}: Arcs between nodes.

S_j::Vector{Int}: Number of states.

I_j::Vector{Vector{Int}}: Information set.

DecisionProgramming.DecisionGraph — Method

Construct and validate a decision graph.

DecisionProgramming.Params — Type

Model parameters.

Arguments

X: Probabilities, X[j][sI(j);sj], ∀j∈C

Y: Consequences, Y[j][s_I(j)], ∀j∈V

U: Utilities map consequences to real valued outcomes.

DecisionProgramming.Params — Method

Construct and validate model parameters.

DecisionProgramming.DecisionModel — Type

Defines the DecisionModel type.

DecisionProgramming.DecisionModel — Method

Initializes the DecisionModel.