API Reference

Model

DecisionProgramming.paths — Function

Iterate over paths.

Iterate over paths with fixed states.

DecisionProgramming.InfluenceDiagram — Type

Influence diagram.

DecisionProgramming.InfluenceDiagram — Method

Construct and validate an influence diagram.

Arguments

C::Vector{Int}: Change nodes.

D::Vector{Int}: Decision nodes.

V::Vector{Int}: Value nodes.

A::Vector{Pair{Int, Int}}: Arcs between nodes.

S_j::Vector{Int}: Number of states.

DecisionProgramming.Params — Type

Decision model parameters.

DecisionProgramming.Params — Method

Construct and validate decision model parameters.

Arguments

diagram::InfluenceDiagram: The influence diagram associated with the probabilities and consequences.

X::Dict{Int, Array{Float64}}: Probabilities

Y::Dict{Int, Array{Float64}}: Consequences

DecisionProgramming.DecisionModel — Type

Defines the DecisionModel type.

DecisionProgramming.DecisionModel — Method

Construct a DecisionModel from an influence diagram and parameters.

Arguments

diagram::InfluenceDiagram

params::Params

DecisionProgramming.probability_sum_cut — Function

Adds a probability sum cut to the model as a lazy constraint.

DecisionProgramming.number_of_paths_cut — Function

Adds a number of paths cut to the model as a lazy constraint.

Analysis

DecisionProgramming.active_paths — Function

Generate all active paths from a decision strategy with fixed states.

Generate all active paths from a decision strategy.

DecisionProgramming.is_compatible — Function

Test is path is compatible with a decision strategy.

DecisionProgramming.state_probabilities — Function

State probabilities.

State probabilities.

DecisionProgramming.utility_distribution — Function

The probability distribution of utilities.

using Plots
x, y = distribution(z, diagram, params)
y2 = cumsum(y)
p = plot(x, y,
    linestyle=:dash,
    markershape=:circle,
    ylims=(0, 1.1),
    label="Distribution",
    legend=:topleft)
plot!(p, x, y2,
    linestyle=:dash,
    markershape=:circle,
    label="Cumulative distribution")
savefig(p, "distribution.svg")

Printing

DecisionProgramming.print_results — Function

Print number of paths, number of active paths and expected utility.

DecisionProgramming.print_decision_strategy — Function

Print decision strategy.

DecisionProgramming.print_state_probabilities — Function

Print state probabilities with fixed states.

Print state probabilities.

Random

DecisionProgramming.random_influence_diagram — Function

Create random influence diagram of given size.

DecisionProgramming.random_params — Function

Generate random params.