API Reference

DecisionProgramming.jl API reference.

model.jl

Influence Diagram

DecisionProgramming.Node — Type

Node type.

DecisionProgramming.State — Type

State type.

DecisionProgramming.InfluenceDiagram — Type

Influence diagram.

DecisionProgramming.InfluenceDiagram — Method

Construct and validate an influence diagram.

Examples

C = [1, 3]
D = [2, 4]
V = [5]
A = [1=>2, 3=>4, 2=>5, 3=>5]
S_j = [2, 3, 2, 4]
G = InfluenceDiagram(C, D, V, A, S_j)

Paths

DecisionProgramming.Path — Type

Path type.

DecisionProgramming.paths — Function

Iterate over paths in lexicographical order.

Examples

julia> collect(paths([2, 3]))[:]
[(1, 1), (2, 1), (1, 2), (2, 2), (1, 3), (2, 3)]

Iterate over paths with fixed states in lexicographical order.

Examples

julia> collect(paths([2, 3], fixed=Dict(1=>2)))[:]
[(2, 1), (2, 2), (2, 3)]

Probabilities

DecisionProgramming.Probability — Type

Type alias for probability.

DecisionProgramming.Probabilities — Type

Probabilities type.

DecisionProgramming.Probabilities — Method

Validate probabilities on an influence diagram.

Examples

X = Probabilities(G, X)

Consequences

DecisionProgramming.Consequence — Type

Type alias for consequence.

DecisionProgramming.Consequences — Type

Consequences type.

DecisionProgramming.Consequences — Method

Validate consequences on an influence diagram.

Examples

Y = Consequences(G, Y)

Path Probability

DecisionProgramming.PathProbability — Type

Path probability type.

Examples

P = PathProbability(G, X)

DecisionProgramming.PathProbability — Method

Evaluate path probability.

Examples

P(s)

Path Utility

DecisionProgramming.PathUtility — Type

Path utility type.

Examples

U = PathUtility(G, Y)

DecisionProgramming.PathUtility — Method

Evaluate path utility.

Examples

U(s)

DecisionProgramming.positive_affine — Method

Evaluate positive affine transformation of the path utility.

Examples

1 ≤ positive_affine(U, s) ≤ 2

Decision Model

DecisionProgramming.variables — Function

Create a multidimensional array of JuMP variables.

DecisionProgramming.DecisionModel — Type

DecisionModel type.

DecisionProgramming.DecisionModel — Method

Construct a DecisionModel from an influence diagram and probabilities.

Examples

model = DecisionModel(G, P; positive_path_utility=true)

DecisionProgramming.probability_sum_cut — Method

Adds a probability sum cut to the model as a lazy constraint.

Examples

probability_sum_cut(model, P)

DecisionProgramming.number_of_paths_cut — Method

Adds a number of paths cut to the model as a lazy constraint.

Examples

atol = 0.9  # Tolerance to trigger the creation of the lazy cut
number_of_paths_cut(model, G, P; atol=atol)

Objective Functions

DecisionProgramming.expected_value — Method

Expected value objective.

Examples

EV = expected_value(model, G, U)

DecisionProgramming.conditional_value_at_risk — Method

Conditional value-at-risk (CVaR) objective. Also known as Expected Shortfall (ES).

Examples

α = 0.05  # Parameter such that 0 ≤ α ≤ 1
ES = conditional_value_at_risk(model, G, U, α)

Decision Strategy

DecisionProgramming.DecisionStrategy — Type

Decision strategy type.

DecisionProgramming.DecisionStrategy — Method

Extract values for decision variables from a decision model.

Examples

Z = DecisionStrategy(model)

analysis.jl

DecisionProgramming.ActivePaths — Type

Interface for iterating over active paths given influence diagram and decision strategy.

Initialize path s of length n

Fill chance states s[C] by generating subpaths paths(G.C)

Fill decision states s[D] by decision strategy Z and path s

Examples

for s in ActivePaths(G, Z)
    ...
end

DecisionProgramming.UtilityDistribution — Type

UtilityDistribution type.

DecisionProgramming.UtilityDistribution — Method

Constructs the probability mass function for path utilities on active paths.

Examples

UtilityDistribution(G, P, U, Z)

DecisionProgramming.StateProbabilities — Type

StateProbabilities type.

DecisionProgramming.StateProbabilities — Method

Associates each node with array of probabilities for each of its states occuring in active paths.

Examples

StateProbabilities(G, P, Z)

DecisionProgramming.StateProbabilities — Method

Associates each node with array of conditional probabilities for each of its states occuring in active paths given fixed states and prior probability.

Examples

# Prior probabilities
prev = StateProbabilities(G, P, Z)

# Select node and fix its state
node = 1
state = 2
StateProbabilities(G, P, Z, node, state, prev)

printing.jl

DecisionProgramming.print_decision_strategy — Function

Print decision strategy.

Examples

print_decision_strategy(G, Z)

DecisionProgramming.print_utility_distribution — Function

Print utility distribution

Examples

udist = UtilityDistribution(G, X, Z)
print_utility_distribution(udist)

DecisionProgramming.print_state_probabilities — Function

Print state probabilities with fixed states.

Examples

sprobs = StateProbabilities(G, X, Z)
print_state_probabilities(sprobs, G.C)
print_state_probabilities(sprobs, G.D)

DecisionProgramming.print_statistics — Function

Print statistics.

DecisionProgramming.print_risk_measures — Function

Print risk measures.

DecisionProgramming.value_at_risk — Method

Value-at-risk.

DecisionProgramming.conditional_value_at_risk — Method

Conditional value-at-risk.

random.jl

DecisionProgramming.InfluenceDiagram — Method

Create random influence diagram of given size.

Examples

import Random
rng = MersenneTwister(1)
n_C, n_D, n_V = 4, 2, 1  # Number of chance, decision and value states
n_A = 2  # Upper limit on information set size for any node
num_states = [2, 3]  # Set of possible number of states
G = InfluenceDiagram(rng, n_C, n_D, n_V, n_A, num_states)

DecisionProgramming.Probabilities — Method

Generate random probabilities for an influence diagram.

Examples

import Random
rng = MersenneTwister(1)
X = Probabilities(rng, G)

DecisionProgramming.Consequences — Method

Generate random consequences for an influence diagram.

Examples

import Random
rng = MersenneTwister(1)
Y = Consequences(rng, G; low = -1.0, high = 1.0)